Youtube Video Caching System
This project demonstrates how Distributed Hash Table works. It uses
Berkeley DB for persistence with LRU cache feature implemented on top of
it. It has an inbuilt HTTP web application server with its own
limitations as it does not use Tomcat. It uses Free Pastry ring that
gives a notion of nodes in a cluster. It uses Youtube Apis to fetch data
from youtube.